+++ src/bf/__init__.py
@@ -2,7 +2,7 @@
 
 import os
 
-__version__ = "11.1.1"
+__version__ = "12.0.0"
 
 
 def main() -> None: